cell color

+++ MemoryMatrix/jni/ColorRGB.cpp
@@ -1,3 +1,3 @@
 #include "ColorRGB.h"
 
-ColorRGB::ColorRGB(uint8_t _r, uint8_t _g, uint8_t _b) : r(_r), g(_g), b(_b){}
+ColorRGB::ColorRGB(float _r, float _g, float _b) : r(_r), g(_g), b(_b){}
