+++ MemoryMatrix/jni/ColorRGB.h
@@ -2,15 +2,13 @@
 #ifndef COLOR_RGB_H_
 #define COLOR_RGB_H_
 
-#include <stdint.h>
-
 class ColorRGB
 {
 public:
-  ColorRGB::ColorRGB(uint8_t _r, uint8_t _g, uint8_t _b);
-  uint8_t r;
-  uint8_t g;
-  uint8_t b;
+  ColorRGB::ColorRGB(float _r, float _g, float _b);
+  float r;
+  float g;
+  float b;
 };
 
 #endif
